build sigToolkit
