--- src/else/PCWindow.cpp
@@ -1,13 +1,0 @@
-/*************************************************************************
-	> File Name: PCWindow.cpp
-	> Author: Doni Daniel
-	> Mail: sigboom@163.com
-	> Created Time: 三  8/28 19:05:09 2019
- ************************************************************************/
-
-#include "PCWindow.h"
-
-void PCWindow::drawContents() {
-    GetView()->DrawOn(this);
-}
-
